requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via django
django==3.2.6
    # via
    #   django-annoying
    #   django-cors-headers
    #   django-cron
    #   django-dbbackup
    #   django-extensions
    #   django-filter
    #   django-model-utils
    #   django-taggit
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   tfc-backend (setup.py)
django-annoying==0.10.6
    # via tfc-backend (setup.py)
django-common-helpers==0.9.2
    # via django-cron
django-cors-headers==3.8.0
    # via tfc-backend (setup.py)
django-cron==0.5.1
    # via tfc-backend (setup.py)
django-dbbackup==3.3.0
    # via tfc-backend (setup.py)
django-extensions==3.1.3
    # via tfc-backend (setup.py)
django-filter==2.4.0
    # via tfc-backend (setup.py)
django-model-utils==4.1.1
    # via tfc-backend (setup.py)
django-taggit==1.5.1
    # via
    #   django-taggit-serializer
    #   tfc-backend (setup.py)
django-taggit-serializer==0.1.7
    # via tfc-backend (setup.py)
djangorestframework==3.12.4
    # via
    #   djangorestframework-simplejwt
    #   tfc-backend (setup.py)
djangorestframework-simplejwt==4.8.0
    # via tfc-backend (setup.py)
pillow==8.3.1
    # via tfc-backend (setup.py)
psycopg2-binary==2.9.1
    # via tfc-backend (setup.py)
pyjwt==2.1.0
    # via djangorestframework-simplejwt
pytz==2021.1
    # via
    #   django
    #   django-dbbackup
rstr==3.0.0
    # via tfc-backend (setup.py)
six==1.16.0
    # via
    #   django-annoying
    #   django-dbbackup
    #   django-taggit-serializer
sqlparse==0.4.1
    # via django
yapf==0.31.0
    # via tfc-backend (setup.py)
